Remove the social_mailer and adjust the social_swiftmail to include the new symfony mailer make it easier to upgrade
