--- modules/social_features/social_mailer/config/schema/social_mailer.schema.yml
@@ -1,30 +1,0 @@
-social_mailer.settings:
-  type: config_object
-  label: 'Social Mailer Settings'
-  mapping:
-    remove_open_social_branding:
-      type: boolean
-      label: 'Open Social Branding will be replaced by site name (and slogan if available).'
-    template_header:
-      type: mapping
-      label: 'Template header'
-      mapping:
-        value:
-          type: text
-          label: 'Header markup'
-        format:
-          type: string
-          label: 'Text format'
-    template_footer:
-      type: mapping
-      label: 'Template footer'
-      mapping:
-        value:
-          type: text
-          label: 'Footer markup'
-        format:
-          type: string
-          label: 'Text format'
-    do_not_send_emails_new_users:
-      type: boolean
-      label: "Don't send email notifications to users who have never logged in"
