--- modules/social_features/social_mailer/social_mailer.info.yml
@@ -1,9 +1,0 @@
-name: 'Social Mailer'
-type: module
-description: 'Social Mailer implementation'
-core_version_requirement: ^9 || ^10
-package: Social
-dependencies:
-  - symfony_mailer:symfony_mailer
-  - activity_send:activity_send_email
-configure: social_mailer.settings
