--- modules/social_features/social_mailer/social_mailer.links.menu.yml
@@ -1,6 +1,0 @@
-social_mailer.settings:
-  title: 'Email notification settings'
-  description: 'E-mail template settings.'
-  route_name: social_mailer.settings
-  parent: social_core.admin.config.social
-  weight: 52
