--- modules/social_features/social_mailer/social_mailer.links.task.yml
@@ -1,4 +1,0 @@
-social_mailer.test:
-  route_name: social_mailer.test
-  title: Test
-  base_route: entity.mailer_transport.collection
