--- modules/social_features/social_mailer/social_mailer.permissions.yml
@@ -1,2 +1,0 @@
-administer social mail:
-  title: 'Administer social mail settings'
