--- modules/social_features/social_mailer/social_mailer.routing.yml
@@ -1,14 +1,0 @@
-social_mailer.settings:
-  path: '/admin/config/opensocial/mail'
-  defaults:
-    _form: 'Drupal\social_mailer\Form\SettingsForm'
-    _title: 'Email notification settings'
-  requirements:
-    _permission: 'administer social mail'
-social_mailer.test:
-  path: '/admin/config/system/mailer/test'
-  defaults:
-    _form: '\Drupal\social_mailer\Form\TestForm'
-    _title: 'Test'
-  requirements:
-    _permission: 'administer social mail'
