+++ modules/social_features/social_swiftmail/social_swiftmail.info.yml
@@ -4,7 +4,8 @@ description: Social Swiftmailer implementation
 core_version_requirement: ^9 || ^10
 package: Social
 dependencies:
-  - swiftmailer:swiftmailer
+  - symfony_mailer:symfony_mailer
+  - activity_send:activity_send_email
   - mailsystem:mailsystem
   - social:social_editor
   - social:social_core
