+++ modules/social_features/social_swiftmail/social_swiftmail.install
@@ -12,26 +12,6 @@
  */
 function social_swiftmail_install() {
   $config_factory = \Drupal::configFactory();
-  // Alter swift settings.
-  $config_factory->getEditable('swiftmailer.message')
-    ->set('format', 'text/html')
-    ->set('respect_format', FALSE)
-    ->save();
-  // Alter mailsystem settings.
-  $config_factory->getEditable('mailsystem.settings')
-    ->set('theme', 'default')
-    ->set('defaults.sender', 'social_swiftmailer')
-    ->set('defaults.formatter', 'social_swiftmailer')
-    ->set('modules.swiftmailer.none.sender', 'social_swiftmailer')
-    ->set('modules.swiftmailer.none.formatter', 'social_swiftmailer')
-    ->save();
-
-  // Ensure the sendmail mode matches the PHP.ini configuration.
-  $sendmail_path = ini_get('sendmail_path');
-  $sendmail_mode = ($sendmail_path !== FALSE && str_contains($sendmail_path, '-bs')) ? 'bs' : 't';
-  $config_factory->getEditable('swiftmailer.transport')
-    ->set('sendmail_mode', $sendmail_mode)
-    ->save();
 
   // Grant the default permissions for this feature.
   user_role_grant_permissions(
@@ -53,6 +33,23 @@ function social_swiftmail_install() {
       'administer social swiftmail',
     ]
   );
+
+  // Remove unnecessary mailer policies and set the correct theme to the default
+  // policy.
+  $keys = $config_factory->listAll('symfony_mailer.mailer_policy.');
+
+  foreach ($keys as $key) {
+    $config = $config_factory->getEditable($key);
+
+    if ($key === 'symfony_mailer.mailer_policy._') {
+      $config
+        ->set('configuration.email_theme.theme', '_default')
+        ->save();
+    }
+    else {
+      $config->delete();
+    }
+  }
 }
 
 /**
