+++ modules/social_features/social_swiftmail/src/Element/SocialSwiftmailTextFormat.php
@@ -29,7 +29,7 @@ class SocialSwiftmailTextFormat {
    *
    * @see social_swiftmail_element_info_alter()
    */
-  public static function process(array &$element, FormStateInterface $form_state, array &$complete_form) {
+  public static function process(array &$element, FormStateInterface $form_state, array &$complete_form): array {
     if (!$form_state->has('views_bulk_operations')) {
       $format = &$element['format'];
