+++ social.info.yml
@@ -20,7 +20,7 @@ dependencies:
   - social:social_activity
   - social:social_core
   - social:social_group
-  - social:social_mailer
+  - social:social_swiftmail
   - social:social_user
 
 # Modules that are installed by default but can be uninstalled.
